+++ tastyworks/models/option_leg.py
@@ -1,0 +1,5 @@
+from tastyworks.models.model import Model
+
+
+class OptionLeg(Model):
+    pass