+++ CHANGELOG.md
@@ -4,6 +4,15 @@ All important and notable changes will be documented here.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.1.0] - 2017-07-26
+
+### Added
+- OrderDetails object to encapsulate order-specific details
+- OptionLeg object to store and process information about option legs
+
+### Changed
+- Internal changes and refactors to the orders model
+- Re-architecture of the 
 
 ## [1.0.0] - 2017-07-24
 *NB: This is backwards-incompatible*